Fix test structure
